include/aon/tools/gui-image-generator/requirements.txt
Pillow==10.3.0
yapf=0.32.0
